requirements.txt
slack_bolt>=1.18.0
gspread>=6.0.0
google-auth>=2.28.0
google-genai>=1.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
